Move lowering analyses to their own directory.
