+++ CMakeLists.txt
@@ -86,7 +86,6 @@ list(APPEND NVFUSER_SRCS
     ${NVFUSER_SRCS_DIR}/graph_fuser.cpp
     ${NVFUSER_SRCS_DIR}/grouped_reduction.cpp
     ${NVFUSER_SRCS_DIR}/index_compute.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/index_compute.cpp
     ${NVFUSER_SRCS_DIR}/instrumentation.cpp
     ${NVFUSER_SRCS_DIR}/ir_base_nodes.cpp
     ${NVFUSER_SRCS_DIR}/ir_builder.cpp
@@ -105,10 +104,17 @@ list(APPEND NVFUSER_SRCS
     ${NVFUSER_SRCS_DIR}/kernel_ir_dispatch.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/alias_memory.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/allocation.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/index_compute.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/divisible_split.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/fused_reduction.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/predicate_elimination.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/shift.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/sync_information.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/thread_predicate.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/trivial_broadcast.cpp
+    ${NVFUSER_SRCS_DIR}/device_lower/analysis/bank_conflict.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/double_buffer.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/divisible_split.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/expr_sort.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/fused_reduction.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/fusion_simplifier.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/index.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/scalar_hoist.cpp
@@ -119,19 +125,13 @@ list(APPEND NVFUSER_SRCS
     ${NVFUSER_SRCS_DIR}/device_lower/magic_zero.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/misaligned_vectorization.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/predicate.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/predicate_elimination.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/replace_size.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/shift.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/sync_information.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/thread_predicate.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/trivial_broadcast.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/unroll.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/utils.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/validation.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/vectorize_welford.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/warp_reduce.cpp
     ${NVFUSER_SRCS_DIR}/device_lower/lower2device.cpp
-    ${NVFUSER_SRCS_DIR}/device_lower/bank_conflict.cpp
     ${NVFUSER_SRCS_DIR}/manager.cpp
     ${NVFUSER_SRCS_DIR}/maxinfo_propagator.cpp
     ${NVFUSER_SRCS_DIR}/multidevice/aggregate_dag.cpp
