+++ csrc/compute_at_map.h
@@ -7,7 +7,7 @@
 // clang-format on
 #pragma once
 
-#include <device_lower/trivial_broadcast.h>
+#include <device_lower/analysis/trivial_broadcast.h>
 #include <disjoint_set.h>
 #include <ir_all_nodes.h>
 #include <kernel_ir.h>
