+++ csrc/device_lower/analysis/bank_conflict.cpp
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
-#include <device_lower/bank_conflict.h>
+#include <device_lower/analysis/bank_conflict.h>
 
 #include <dynamic_type.h>
 #include <expr_evaluator.h>
