+++ csrc/device_lower/analysis/index_compute.cpp
@@ -6,7 +6,7 @@
  */
 // clang-format on
 #include <contiguity.h>
-#include <device_lower/index_compute.h>
+#include <device_lower/analysis/index_compute.h>
 #include <device_lower/lower2device.h>
 #include <device_lower/magic_zero.h>
 #include <device_lower/utils.h>
