+++ csrc/device_lower/analysis/thread_predicate.cpp
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
-#include <device_lower/thread_predicate.h>
+#include <device_lower/analysis/thread_predicate.h>
 
 #include <device_lower/lower2device.h>
 #include <device_lower/utils.h>
