+++ csrc/device_lower/magic_zero.cpp
@@ -7,7 +7,7 @@
 // clang-format on
 #include <device_lower/magic_zero.h>
 
-#include <device_lower/index_compute.h>
+#include <device_lower/analysis/index_compute.h>
 #include <device_lower/lower2device.h>
 #include <dispatch.h>
 #include <instrumentation.h>
