+++ csrc/device_lower/utils.cpp
@@ -9,8 +9,8 @@
 
 #include <ATen/cuda/CUDAContext.h>
 #include <c10/util/irange.h>
+#include <device_lower/analysis/thread_predicate.h>
 #include <device_lower/lower2device.h>
-#include <device_lower/thread_predicate.h>
 #include <ir_iostream.h>
 #include <ir_utils.h>
 #include <iter_visitor.h>
