+++ csrc/fusion.cpp
@@ -6,7 +6,7 @@
  */
 // clang-format on
 #include <codegen.h>
-#include <device_lower/bank_conflict.h>
+#include <device_lower/analysis/bank_conflict.h>
 #include <device_lower/lower2device.h>
 #include <disjoint_set.h>
 #include <executor_params.h>
