+++ csrc/scheduler/vectorize_helper.cpp
@@ -9,7 +9,7 @@
 
 #include <compute_at_map.h>
 #include <contiguity.h>
-#include <device_lower/divisible_split.h>
+#include <device_lower/analysis/divisible_split.h>
 #include <expr_evaluator.h>
 #include <ir_builder.h>
 #include <iter_visitor.h>
